+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from "./lib/DeleteAction/DeleteAction";
-export * from "./lib/Utils/DateHandler";
 export * from "./lib/Layout/AuthenticationLayout";
 
 // Panel sidebar layout