+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `DateHandler` utils as it belongs to another package and will be replaced in the future
+
 ### Fixed
 
 - Fixed PanelSideBarLayout topBarCustomItems displaying