+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PanelSidebar cypress test and toggle it on any screen size
 
+### Removed
+
+- `DateHandler` utils as it belongs to another package and will be replaced in the future
+
 ### Fixed
 
 - Fixed PanelSideBarLayout topBarCustomItems displaying