+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- the `DeleteAction` component now correctly displays the delete button, and supports custom text.
+
 ## [2.8.0] - 2023-09-05
 
 ### Fixed