+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.8.1] - 2023-09-19
+
 ### Fixed
 
 - the `DeleteAction` component now correctly displays the delete button, and supports custom text.
@@ -113,12 +115,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - created package :tada:
 - added tests
 
-[unreleased]: https://github.com/neolution-ch/react-pattern-ui/compare/2.8.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-pattern-ui/compare/2.8.1...HEAD
 [2.1.1]: https://github.com/neolution-ch/react-pattern-ui/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/neolution-ch/react-pattern-ui/compare/2.0.2...2.1.0
 [2.0.2]: https://github.com/neolution-ch/react-pattern-ui/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/neolution-ch/react-pattern-ui/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/neolution-ch/react-pattern-ui/compare/45cbeb76034667019da84605082679900f506d75...2.0.0
+[2.8.1]: https://github.com/neolution-ch/react-pattern-ui/compare/2.8.0...2.8.1
 [2.8.0]: https://github.com/neolution-ch/react-pattern-ui/compare/2.7.0...2.8.0
 [2.7.0]: https://github.com/neolution-ch/react-pattern-ui/compare/2.6.0...2.7.0
 [2.6.0]: https://github.com/neolution-ch/react-pattern-ui/compare/2.5.0...2.6.0